# docs/swagger.yaml
...
info:
  contact: {}
paths:
...
  /api/v1/subscription/:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 0, "data": []}'
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      summary: 订阅列表
      tags:
      - 订阅管理
  /api/v1/subscription/{id}/:
    get:
      consumes:
      - application/json
      parameters:
      - description: 订阅id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 0, "data":{}}'
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      summary: 查看订阅
      tags:
      - 订阅管理
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"